requirements.txt

Django>=4.2
djangorestframework
djangorestframework-simplejwt
Pillow
django-cors-headers
